El Precio Justo - Merge con Haml . Se parsea float para price

+++ sprint/code/leo/precio_justo_service_factory/lib/serviceimplementation/offline/offline_search_strategies.rb
@@ -69,7 +69,7 @@ class OfflineFilterByPrice < OfflineFilter
 
   def initialize value
     super
-    @condition = Proc.new { |x| x.price? <= @value.to_i }
+    @condition = Proc.new { |x| x.price? <= @value.to_f }
   end
 end
